bug report with Mouse not overlapping on DEFEAT
here's a bug report. I create rules as : MOUSE IS YOU, CAT IS DEFEAT . but mouse could push the CAT around that should not happen. You cannot move the Defeat or KILL or MELT or WATER OR SINK objects. Unless the rule is changed for these 3 objects

also can you please add the option for the LAVA sprite to be added.